requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
flask-mysqldb==2.0.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
requests==2.31.0
Werkzeug==3.0.1
Flask-WTF==1.2.1
gunicorn==21.2.0
